// src/app/features/workout/models/workouts-per-month-dto.ts
export type WorkoutsPerMonthDto = {
    years: number[],
    januaryWorkouts: number;
    februaryWorkouts: number;
    marchWorkouts: number;
    aprilWorkouts: number;
    mayWorkouts: number;
    juneWorkouts: number;
    julyWorkouts: number;
    augustWorkouts: number;
    septemberWorkouts: number;
    octoberWorkouts: number;
    novemberWorkouts: number;
    decemberWorkouts: number;
};